Add a graph of n_patches by author and how long it took to merge
x axis is the number of patches by an author.
y axis is how long on average it took to merge his or her patches.

This shows the expected correlation: The more patches someone sends,
the shorter it is to merge his/her patches, but the correlation
only forms after an initial set of patches.